# optimize encoding