the result of `go test -bench . -benchmem -count 10`:

```
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/forwarded-header
         │  .old.txt   │              .new.txt               │
         │   sec/op    │   sec/op     vs base                │
String-8   238.9n ± 2%   178.6n ± 0%  -25.24% (p=0.000 n=10)
Parse-8    555.3n ± 3%   528.6n ± 3%   -4.80% (p=0.000 n=10)
geomean    364.2n        307.3n       -15.64%

         │  .old.txt  │              .new.txt               │
         │    B/op    │    B/op     vs base                 │
String-8   144.0 ± 0%   136.0 ± 0%  -5.56% (p=0.000 n=10)
Parse-8    408.0 ± 0%   408.0 ± 0%       ~ (p=1.000 n=10) ¹
geomean    242.4        235.6       -2.82%
¹ all samples are equal

         │  .old.txt  │               .new.txt               │
         │ allocs/op  │ allocs/op   vs base                  │
String-8   7.000 ± 0%   5.000 ± 0%  -28.57% (p=0.000 n=10)
Parse-8    8.000 ± 0%   8.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean    7.483        6.325       -15.48%
¹ all samples are equal
```